% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_read_redshift.R
\name{spark_read_redshift}
\alias{spark_read_redshift}
\title{SQL Query from Amazon Redshift into a Spark DataFrame}
\usage{
spark_read_redshift(sc, name, query, redshift_url, redshift_user,
  redshift_password, s3bucket, s3accesskey, s3secret)
}
\arguments{
\item{sc}{a spark connection}

\item{name}{the name to assign to the newly generated table}

\item{query}{SQL query to execute on database}

\item{redshift_url}{the url to the redshift database}

\item{redshift_user}{Amazon Redshift Username}

\item{redshift_password}{Amazon Redshift Password}

\item{s3bucket}{AWS S3 bucket address (must have write access)}

\item{s3accesskey}{AWS Access Key ID}

\item{s3secret}{AWS Access Secret}
}
\description{
SQL Query from Amazon Redshift into a Spark DataFrame
}

