% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{sparkredshift}
\alias{sparkredshift}
\alias{sparkredshift-package}
\title{Read data from Amazon Redshift into Spark DataFrames}
\description{
'sparkredshift' is an extension for sparklyr to read data from Amazon Redshift into Spark DataFrames. It uses Spark package spark-redshift to data from Amazon Redshift into Spark DataFrames.
}

